product combiner

+++ products/models.py
@@ -13,6 +13,7 @@ class Manufacturer(models.Model):
 
 class Category(models.Model):
     name = models.CharField('name', max_length=30, blank=True, null=True)
+    is_active = models.BooleanField(default=False)
 
     def __str__(self):
         return "<Category %s>" % self.name
@@ -118,11 +119,6 @@ class Product(models.Model):
             if price:
                 prices.append(price)
 
-        self.price = min(prices) if prices else None
-        self.manufacturing_name = most_frequent(manufacturing_names)
-        self.update_name(names)
-        self.update_specs(specs_list)
-
         # Update Category
         category_name = most_frequent(categories)
         if category_name:
@@ -132,15 +128,21 @@ class Product(models.Model):
             if self.category and self.category.products.count() <= 1: self.category.delete()
             self.category = category
 
-        # Update Manufacturer
-        first_names = [name.split(' ', 1)[0] for name in names]
-        manufacturer_name = most_frequent(first_names)
-        if manufacturer_name:
-            try: manufacturer = Manufacturer.objects.get(name=manufacturer_name)
-            except: manufacturer = Manufacturer.objects.create(name=manufacturer_name)
+        if self.category.is_active:
+            self.price = min(prices) if prices else None
+            self.manufacturing_name = most_frequent(manufacturing_names)
+            self.update_name(names)
+            self.update_specs(specs_list)
 
-            if self.manufacturer and self.manufacturer.products.count() <= 1: self.manufacturer.delete()
-            self.manufacturer = manufacturer
+            # Update Manufacturer
+            first_names = [name.split(' ', 1)[0] for name in names]
+            manufacturer_name = most_frequent(first_names)
+            if manufacturer_name:
+                try: manufacturer = Manufacturer.objects.get(name=manufacturer_name)
+                except: manufacturer = Manufacturer.objects.create(name=manufacturer_name)
+
+                if self.manufacturer and self.manufacturer.products.count() <= 1: self.manufacturer.delete()
+                self.manufacturer = manufacturer
 
     def get_price(self):
         return min([mp.get_price() for mp in self.meta_products.all()])
